% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concstats.R
\name{concstats}
\alias{concstats}
\title{Market structure and concentration measures.}
\usage{
concstats(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{logical vector that indicates whether NA values should be excluded}
}
\value{
returns a vector of calculated measures in table format.
}
\description{
A convenience function which calculates a set of different
market structure and concentration measures, e.g. k-firm ratios, HHI, dual of HHI,
Stenbacka index, Entropy and others.
}
\note{
the vector of market shares should be in a decimal format corresponding to total
share of indivudual firms/units (e.g. share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04 )).
The sum of the vector/share should sum up to 1. The results are presented in a table
}
\examples{
share <- c(0.4, 0.266, 0.334, 0, 0) # a vector of market shares
con_share <- concstats(share)  # concentration measures of the vector

}
