% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shares.R
\name{concstats_shares}
\alias{concstats_shares}
\title{shares}
\usage{
concstats_shares(x, na.rm = TRUE, digits = NULL)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}. If set to \code{FALSE} the computation yields a message
if the vector contains \code{NA} values. NAs will be removed for further
computations.}

\item{digits}{An optional value for digits. Specifies the minimum number of
significant digits to be printed in values. The default is \code{NULL} and
will use base R print option.}
}
\value{
A numeric vector in decimal form.
}
\description{
The \code{concstats_shares} function is a helper function making it easier
to convert numeric variable into individual shares. This might be convenient
for larger vectors.
}
\details{
\code{concstats_shares} is a helper function. The user can manually
convert or provide numerical vectors of shares or use
\code{constats_shares}.
}
\examples{
# a vector of loans (without special characters, e.g. currency symbols)
x <- c(538572286.08, 481096.77, 161914143.03, 128796268.59, 69055940.72)
concstats_shares(x)

# a vector with NA values
x2 <- c(538572286.08, 481096.77, 161914143.03, 128796268.59, 69055940.72, NA)
concstats_shares(x2, na.rm = TRUE, digits = 3)

}
