% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5_df.R
\name{concstats_top5_df}
\alias{concstats_top5_df}
\title{Top 5 market shares data frame}
\usage{
concstats_top5_df(x, y, digits = NULL)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{y}{A non-negative vector of shares. All integers (e.g. sales) are
converted to relative decimal numbers.}

\item{digits}{An optional value for digits. Specifies the minimum number of
significant digits to be printed in values. The default is \code{NULL} and
will use base R print option. Significant digits defaults to 7. Values are
restricted between 1 and default value.}
}
\value{
A \verb{data frame}.
}
\description{
Top 5 market shares data frame
}
\note{
Note that the first column in your data frame will be the index
and values should be unique.
}
\examples{
x <- data.frame(
firm = c("A", "B", "C", "D", "E"),
share = c(0.2,0.25,0.1,0.05,0.4)
)
concstats_top5_df(x, "share", digits = 2)

}
\seealso{
Other Market structure measures:
\code{\link[=concstats_all_mstruct]{concstats_all_mstruct()}},
\code{\link[=concstats_firm]{concstats_firm()}},
\code{\link[=concstats_mstruct]{concstats_mstruct()}},
\code{\link[=concstats_nrs_eq]{concstats_nrs_eq()}},
\code{\link[=concstats_top]{concstats_top()}},
\code{\link[=concstats_top3]{concstats_top3()}},
\code{\link[=concstats_top3_df]{concstats_top3_df()}},
\code{\link[=concstats_top5]{concstats_top5()}},
\code{\link[=concstats_top_df]{concstats_top_df()}}
}
\concept{Market structure measures}
