% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{share_2016_total}
\alias{share_2016_total}
\title{Total market share}
\format{
A data frame with 22 rows and 3 variables: \code{coop_id}, \code{total} and
\code{share}

@source Author
}
\usage{
share_2016_total
}
\description{
A data set with the total market share for each credit cooperative
}
\keyword{datasets}
