% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{concstats_comp}
\alias{concstats_comp}
\alias{concstats_hhi}
\alias{concstats_hhi_min}
\alias{concstats_hhi_d}
\alias{concstats_dom}
\alias{concstats_sten}
\alias{concstats_all_comp}
\title{Concentration Measures}
\usage{
concstats_comp(x, unbiased = FALSE, type = c("hhi", "hhi_d", "hhi_min",
 "dom", "sten", "all"),na.rm = TRUE)

concstats_hhi(x, unbiased = FALSE, na.rm = TRUE)

concstats_hhi_min(x, na.rm = TRUE)

concstats_hhi_d(x, na.rm = TRUE)

concstats_dom(x, na.rm = TRUE)

concstats_sten(x, na.rm = TRUE)

concstats_all_comp(x, na.rm = TRUE)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{unbiased}{Logical. Argument specifying whether or not a finite sample
correction should be applied. Must be either TRUE or FALSE. The default is
FALSE.}

\item{type}{A character string of the measure to be calculated,
can be abbreviated with the first letter. Defaults to "hhi". Input is not
case-sensitive.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either TRUE or FALSE. The default is FALSE.}
}
\value{
prints the calculated measure.

a data.frame of concentration and competition measures with default
settings.
}
\description{
A set of different concentration measures.
}
\details{
\code{concstats_comp} is a wrapper for the proposed concentration measures
\code{concstats_hhi}, \code{concstats_hhi_d}, \code{concstats_hhi_min},
\code{concstats_dom}, \code{concstats_sten}, \code{concstats_all}.
If no measure is specified "hhi" will be the default.

\code{concstats_hhi}, \code{concstats_hhi_min}, \code{concstats_hhi_d}
calculate the Herfindahl-Hirschman index, its minimum, and its dual,
respectively. \code{concstats_dom} calculates a dominance index and
\code{concstats_sten} calculates the Stenbacka index. The index indicates
the market share of a dominant position.

All measures can be accessed individually.
\code{concstats_hhi}, \code{concstats_hhi_d}, and \code{concstats_dom} can
be calculated individually as a normalized measure changing the default
setting to \code{TRUE}. \code{concstats_all} computes all measures in a one
step procedure.
}
\note{
the vector of market shares should be in a decimal form corresponding
to total shares of individual firms/units. The sum of the vector should sum
up to 1.
}
\examples{
# a vector of market shares
x <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the Herfindahl-Hirschman index of the vector
shares_hhi <- concstats_comp(x, type = "hhi")
# individual measure
shares_sten <- concstats_sten(x)
# complete group measures
shares_comp <- concstats_comp(x, type = "all")

}
\seealso{
{\code{\link{concstats_concstats}}, \code{\link{concstats_mstruct}},
\code{\link{concstats_inequ}}}
}
