% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{concstats_comp}
\alias{concstats_comp}
\alias{concstats_hhi}
\alias{concstats_hhi_min}
\alias{concstats_hhi_d}
\alias{concstats_dom}
\alias{concstats_sten}
\alias{concstats_all_comp}
\title{Concentration Measures}
\usage{
concstats_comp(x, normalized = FALSE, type = c("hhi", "hhi_d", "hhi_min",
 "dom", "sten", "all"),na.rm = TRUE)

concstats_hhi(x, normalized = FALSE, na.rm = TRUE)

concstats_hhi_min(x, na.rm = TRUE)

concstats_hhi_d(x, na.rm = TRUE)

concstats_dom(x, na.rm = TRUE)

concstats_sten(x, na.rm = TRUE)

concstats_all_comp(x, na.rm = TRUE)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{normalized}{Logical. Argument specifying whether or not a normalized
value is required. Ranges from {0, 1} and often used for comparison over
time. Must be either TRUE or FALSE. The default is FALSE.}

\item{type}{A character string of the measure to be calculated,
can be abbreviated with the first letter. Defaults to "hhi". Input is not
case-sensitive.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either TRUE or FALSE. The default is FALSE.}
}
\value{
the calculated numeric measure.

\code{data.frame} of concentration and competition measures with default
settings.
}
\description{
A set of different concentration measures.
}
\details{
\code{concstats_comp} is a wrapper for the proposed concentration measures
\code{concstats_hhi}, \code{concstats_hhi_d}, \code{concstats_hhi_min},
\code{concstats_dom}, \code{concstats_sten}, \code{concstats_all}.
If no measure is specified "hhi" will be the default.

\code{concstats_hhi} calculates the widely used Herfindahl-Hirschman Index
(Herfindahl, 1950 and Hirschman, 1945), \code{concstats_hhi_min}
its minimum, and \code{concstats_hhi_d} its dual (Chang, Guerra,
de Souza Penaloza & Tabak, 2005)
\code{concstats_dom} calculates a dominance index (Garcia, 1994) and
\code{concstats_sten} calculates the Stenbacka index (Melnik, Shy, and
Stenbacka, 2005). The index indicates the market share of a dominant
position.

All measures can be accessed individually.
\code{concstats_hhi}, \code{concstats_hhi_d}, and \code{concstats_dom} can
be calculated individually as a normalized measure changing the default
setting to \code{TRUE}. \code{concstats_all} computes all measures in a one
step procedure.
}
\note{
The vector of market shares should be in a decimal form corresponding
to total shares of individual firms/units. The vector should sum up to 1.
}
\examples{
# a vector of market shares
x <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the Herfindahl-Hirschman index of the vector
concstats_comp(x, type = "hhi")
# individual measure
concstats_sten(x)
# complete group measures
concstats_comp(x, type = "all")

}
\references{
Herfindahl, O. C. (1950), "Concentration in the steel industry"
(PhD thesis)

Hirschmann, A. O. (1945), "National power and structure of
foreign trade".

Chang, E. J., Guerra, S. M., de Souza Penaloza, R. A. & Tabak,
B. M. (2005) Banking concentration: the Brazilian case. In Financial
Stability Report. Brasilia: Banco Central do Brasil, 4: 109-129.

Garcia Alba Idunate, P. (1994). Un Indice de dominancia para el
analisis de la estructura de los mercados. El Trimestre Economico,
61: 499-524.

Melnik, A., Shy, Oz, Stenbacka, R., (2008), Assessing market
dominance, Journal of Economic Behavior and Organization, 68: pp. 63-72.
}
\seealso{
\code{\link[concstats_concstats]{concstats_concstats()}},
\code{\link[concstats_mstruct]{concstats_mstruct()}},
\code{\link[concstats_inequ]{concstats_inequ()}}
}
