% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstruct.R
\name{concstats_firm}
\alias{concstats_firm}
\title{Number of firms}
\usage{
concstats_firm(x, na.rm = TRUE)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{na.rm}{Logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}.
If set to \code{FALSE} the computation yields \code{NA} if the vector
contains \code{NA} values.}
}
\value{
A positive numeric integer.
}
\description{
Number of firms
}
\examples{
# a vector of market shares
x <- c(0.4, 0.2, 0.25, 0.1, 0.05)
concstats_firm(x)
}
