% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequ.R
\name{concstats_inequ}
\alias{concstats_inequ}
\alias{concstats_entropy}
\alias{concstats_gini}
\alias{concstats_simpson}
\alias{concstats_palma}
\alias{concstats_grs}
\alias{concstats_all_inequ}
\title{Inequality and Diversity Measures}
\usage{
concstats_inequ(x, unbiased = FALSE, type = c("entropy", "gini", "simpson",
 "palma", "grs", "all"), na.rm = TRUE)

concstats_entropy(x, unbiased = TRUE, na.rm = TRUE)

concstats_gini(x, unbiased = FALSE, na.rm = TRUE)

concstats_simpson(x, unbiased = FALSE, na.rm = TRUE)

concstats_palma(x, na.rm = TRUE)

concstats_grs(x, na.rm = TRUE)

concstats_all_inequ(x, na.rm = TRUE)
}
\arguments{
\item{x}{a non-negative numeric vector.}

\item{unbiased}{Logical. Argument specifying whether or not a finite sample
correction should be applied. Must be either TRUE or FALSE. The default is
FALSE.}

\item{type}{a character string of the measure to be calculated, defaults to
"concstats_entropy". Input is not case-sensitive.}

\item{na.rm}{a logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either TRUE or FALSE. The default is TRUE.
If set to \code{FALSE} the computation yields \code{NA}.}
}
\value{
the calculated numeric measure

a \code{data.frame} of inequality measures with default settings.
}
\description{
A set of different inequality and diversity measures.
}
\details{
\code{concstats_inequ} is a wrapper for the proposed inequality measures
\code{concstats_entropy}, \code{concstats_gini}, \code{concstats_simpson},
\code{concstats_palma}, \code{concstats_grs}, \code{concstats_all}.
If no measure is specified, "concstats_entropy" is the default.

\code{concstats_entropy} returns the (Shannon) Entropy, \code{concstats_gini}
is the Gini coefficient, \code{concstats_simpson} is the complement of the
Herfindahl-Hirschman Index. You can normalize each of these three measures
by setting \code{unbiased = TRUE}

\code{concstats_palma} measures the ratio of inequality (normally used with
income inequality) of the top 10 percent to the bottom 40 percent.
\code{concstats_grs} is an alternative inequality measure (Ginevicius, 2009),
\code{concstats_all} returns all measures in a one step procedure.
}
\note{
the non-negative vector of market shares should be in a decimal form
corresponding to the total shares of individual firms/units.The vector
should sum up to 1.
}
\examples{
# a vector of market shares
share <- c(0.4, 0.2, 0.25, 0.1, 0.05)
# Calculate the Palma ratio
share_p <- concstats_inequ(share, type = "palma")
# Calculate the entropy measure directly
share_ent <- concstats_entropy(share, unbiased = TRUE)
# Calculate the group measures
share_inequ <- concstats_inequ(share, type = "all")

}
\seealso{
{\code{\link{concstats_concstats}}, \code{\link{concstats_mstruct}},
\code{\link{concstats_comp}}}
}
