% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstruct.R
\name{concstats_mstruct}
\alias{concstats_mstruct}
\alias{concstats_firm}
\alias{concstats_nrs_eq}
\alias{concstats_top}
\alias{concstats_top3}
\alias{concstats_top5}
\alias{concstats_all_mstruct}
\title{Market Structure Measures}
\usage{
concstats_mstruct(x, type = c("firm", "nrs_eq", "top", "top3", "top5",
 "all"), na.rm = TRUE)

concstats_firm(x, na.rm = TRUE)

concstats_nrs_eq(x, na.rm = TRUE)

concstats_top(x, na.rm = TRUE)

concstats_top3(x, na.rm = TRUE)

concstats_top5(x, na.rm = TRUE)

concstats_all_mstruct(x, na.rm = TRUE)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{type}{A character string of the measure to be calculated,
can be abbreviated with the first letter. Defaults to "firm". Input is not
case-sensitive.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either TRUE or FALSE. The default is TRUE.
If set to \code{FALSE} the computation yields \code{NA}.}
}
\value{
prints the calculated measure.

a data.frame of market structure measures with default settings.
}
\description{
Set of different market structure measures to reflect a given
market structure.
}
\details{
\code{concstats_mstruct} is a wrapper for the proposed structural measures
\code{concstats_firm}, returns the number of firms with a given market share
\code{concstats_nrs_eq} computes the numbers equivalent, \code{concstats_top},
\code{concstats_top3}, and \code{concstats_top5}calculate the share of the
top (top 3 and top 5) firm(s). \code{concstats_all} computes all measures in
a one step procedure. All measures can be computed individually.
}
\note{
the vector of market shares should be in a decimal form corresponding
to total share of individual firms/units.The sum of the vector should sum up
to 1.
}
\examples{
# a vector of market shares
share <- c(0.35, 0.4, 0.05, 0.1, 0.06, 0.04)
# the number of firms with market share
share_firm <- concstats_mstruct(share, type = "firm")
# Calculate top market share individually
share_top <- concstats_top(share)
# Calculate the market structure group measures
share_mstruct <- concstats_mstruct(share, type = "all")

}
\seealso{
{\code{\link{concstats_concstats}}, \code{\link{concstats_comp}},
\code{\link{concstats_inequ}}}
}
