% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberseq.R
\name{concstats_nrs_eq}
\alias{concstats_nrs_eq}
\title{Numbers equivalent}
\usage{
concstats_nrs_eq(x, na.rm = TRUE, digits = NULL)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}.
If set to \code{FALSE} the computation yields \code{NA} if \code{NA} values
are present.}

\item{digits}{An optional value for digits. Specifies the minimum number of
significant digits to be printed in values. The default is \code{NULL} and
will use base R print option.}
}
\value{
A positive numeric value, rounded to the second decimal place.
}
\description{
Numbers equivalent
}
\examples{
# a vector of market shares
x <- c(0.4, 0.2, 0.25, 0.1, 0.05)
concstats_nrs_eq(x, digits = 2)

}
\seealso{
Other Market structure measures: 
\code{\link{concstats_all_mstruct}()},
\code{\link{concstats_firm}()},
\code{\link{concstats_mstruct}()},
\code{\link{concstats_top}()},
\code{\link{concstats_top3}()},
\code{\link{concstats_top5}()},
\code{\link{concstats_top5_df}()},
\code{\link{concstats_top_df}()}
}
\concept{Market structure measures}
