% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top3.R
\name{concstats_top3}
\alias{concstats_top3}
\title{Top 3 market share}
\usage{
concstats_top3(x, na.rm = TRUE, digits = NULL)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}. If set to \code{FALSE} the computation yields a message
if the vector contains \code{NA} values. NAs will be removed for further
computations.}

\item{digits}{An optional value for digits. Specifies the minimum number of
significant digits to be printed in values. The default is \code{NULL} and
will use base R print option.}
}
\value{
A positive numeric value, which indicates the cumulative sum
of the top 3 market shares as a percentage.
}
\description{
Top 3 market share
}
\examples{
# a vector of market shares
x <- c(0.4, 0.2, 0.25, 0.1, 0.05)
concstats_top3(x, digits = 2)

}
\seealso{
Other Market structure measures: 
\code{\link{concstats_all_mstruct}()},
\code{\link{concstats_firm}()},
\code{\link{concstats_mstruct}()},
\code{\link{concstats_nrs_eq}()},
\code{\link{concstats_top}()},
\code{\link{concstats_top3_df}()},
\code{\link{concstats_top5}()},
\code{\link{concstats_top5_df}()},
\code{\link{concstats_top_df}()}
}
\concept{Market structure measures}
