% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstruct.R
\name{concstats_top3}
\alias{concstats_top3}
\title{Top 3 market share}
\usage{
concstats_top3(x, na.rm = TRUE)
}
\arguments{
\item{x}{A non-negative numeric vector.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}.
If set to \code{FALSE} the computation yields \code{NA} if the vector
contains \code{NA} values.}
}
\value{
A positive numeric value, which indicates the sum of the top 3
market shares as a percentage.
}
\description{
Top 3 market share
}
\examples{
# a vector of market shares
x <- c(0.4, 0.2, 0.25, 0.1, 0.05)
concstats_top3(x)
}
