% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top3_df.R
\name{concstats_top3_df}
\alias{concstats_top3_df}
\title{Top market share data frame}
\usage{
concstats_top3_df(x, y, digits = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A non-negative vector of shares. All integers (e.g. sales) are
converted to relative decimal numbers.}

\item{digits}{An optional value for digits. Specifies the minimum number of
significant digits to be printed in values. The default is \code{NULL} and
will use base R print option. Significant digits defaults to 7. Values are
restricted between 1 and default value.}
}
\description{
Top market share data frame
}
\examples{
x <- data.frame(
firm = c("A", "B", "C", "D", "E"),
share = c(0.2,0.25,0.1,0.05,0.4)
)
concstats_top3_df(x, "share", digits = 2)

}
