% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_df.R
\name{concstats_top_df}
\alias{concstats_top_df}
\title{Top market share data frame}
\usage{
concstats_top_df(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A non-negative vector of shares.}

\item{na.rm}{A logical vector that indicates whether \code{NA} values should
be excluded or not. Must be either \code{TRUE} or \code{FALSE}. The default
is \code{TRUE}.
If set to \code{FALSE} the computation yields \code{NA} if \code{NA} values
are present.}
}
\value{
A data frame, which indicates an id or firm column and the top market
share in decimal form.
}
\description{
Top market share data frame
}
\examples{
# some data
id <- c(1, 2, 3, 4, 5)
x <- c(0.2, 0.25, 0.4, 0.1, 0.05)
test_df <- data.frame(id, x)

concstats_top_df(test_df, "x")


}
\seealso{
Other Market structure measures: 
\code{\link{concstats_all_mstruct}()},
\code{\link{concstats_firm}()},
\code{\link{concstats_mstruct}()},
\code{\link{concstats_nrs_eq}()},
\code{\link{concstats_top}()},
\code{\link{concstats_top3}()},
\code{\link{concstats_top3_df}()},
\code{\link{concstats_top5}()},
\code{\link{concstats_top5_df}()}
}
\concept{Market structure measures}
